[CONTINUUM-2052] added confirmation pages when deleting installations and build definition templates
Submitted By: Jan Ancajas


git-svn-id: https://svn.eu.apache.org/repos/asf/continuum/trunk@752008 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-webapp/src/main/java/org/apache/maven/continuum/web/action/admin/BuildDefinitionTemplateAction.java
@@ -153,9 +153,17 @@ public class BuildDefinitionTemplateAction
     public String delete()
         throws BuildDefinitionServiceException
     {
-        buildDefinitionTemplate = getContinuum().getBuildDefinitionService()
-            .getBuildDefinitionTemplate( this.buildDefinitionTemplate.getId() );
-        this.getContinuum().getBuildDefinitionService().removeBuildDefinitionTemplate( buildDefinitionTemplate );
+        if ( confirmed )
+        {
+            buildDefinitionTemplate =
+                getContinuum().getBuildDefinitionService().getBuildDefinitionTemplate(
+                                                                                       this.buildDefinitionTemplate.getId() );
+            this.getContinuum().getBuildDefinitionService().removeBuildDefinitionTemplate( buildDefinitionTemplate );
+        }
+        else
+        {
+            return CONFIRM;
+        }
         return SUCCESS;
     }
     
@@ -233,8 +241,16 @@ public class BuildDefinitionTemplateAction
     public String deleteBuildDefinition()
         throws BuildDefinitionServiceException
     {
-        buildDefinition = getContinuum().getBuildDefinitionService().getBuildDefinition( this.buildDefinition.getId() );
-        this.getContinuum().getBuildDefinitionService().removeBuildDefinition( buildDefinition );
+        if ( confirmed )
+        {
+            buildDefinition = getContinuum().getBuildDefinitionService().getBuildDefinition( this.buildDefinition.getId() );
+            this.getContinuum().getBuildDefinitionService().removeBuildDefinition( buildDefinition );
+        }
+        else
+        {
+            return CONFIRM;
+        }
+        
         return SUCCESS;
     }
